requirements.txt
# Core ML / NLP
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4

# LLM APIs
openai>=1.30.0
groq>=0.9.0

# Document parsing
pdfplumber>=0.10.0
unstructured>=0.13.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# RAG / retrieval
rank-bm25>=0.2.2
tiktoken>=0.7.0

# Evaluation
ragas>=0.1.9
datasets>=2.19.0

# Data / utilities
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.7.0
rich>=13.7.0

# Web app
gradio>=4.31.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
